Update New Year's.Rmd
Update of the New Year's eve Markdown file. Nishtha wrote the code which included graphs and tables. I did the final check (a new devide which didn't know the datasets yet) and everything worked. The following seps are to try to include more loops and see if we are missing anything else.